Implement JoinAction to allow no quotation mark on ordered arguments
OVERVIEW:

This means that commands like:
$ mal list plan to watch
$ mal search serial experiments lain
$ mal list --user lubien on hold

Can be possible, where before I need to do:

$ mal list 'plan to watch'
$ mal search 'serial experiments lain'
$ mal list --user lubien 'on hold'

IMPLICATIONS:

However, I need to break-up the list choices on just tokens, because
the choice checking happens before the JoinAction.__call__. So if
I not did that, the mal list will not work when `plan to watch` and
'on hold' would be called. This can be more confusing when getting
error and will return just the tokens which you can compose the
section. And this will worse because things like "plan to hold" will
be possible, will pass by parsing args, but nothing will be returned
(silently error). Maybe the docs can clarify this,
but I'm wondering if I'll add that really (unlikely).

BTW, increase/decrease command cannot be used with this, because they
have an optional argument after the anime_regex, so this doesn't works
nicely. When some user call this:

$ mal inc serial experiments 4

There is no way to look what is the anime_regex and the episode
argument. Because the animex_regex before get all the next arguments passed.
So, this will be passed as anime_regex='serial experiments 4', which
indeed is not the behavior expected. So I didn't even add the new action
class for that subcommand.

Well, after wrote this summary shit, this seems very bad.
Maybe I'll just drop later.