Remove JoinAction to list argument: too messy
Beyond the problem I described on the last commit, the old
functionality is dropped. This is not good. At least on another
commands the user can still using quotation marks and he will be
fine. But not here, each individual token should be in the
choices lists and 'plan to watch' with quotation is treated
as a individual token.

So... removing that. I don't want make the software more complex.